requirements.txt
flask
opencv-python
numpy
requests
paho-mqtt==1.6.1